% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/later2test-package.R
\docType{package}
\name{later2test-package}
\alias{later2test}
\alias{later2test-package}
\title{later2test: Additional Unit Tests for 'later'}
\description{
Unit tests that use the 'later2' API to test the behaviour of 'later2' itself. This package is not intended for general use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://later.r-lib.org}
  \item \url{https://github.com/r-lib/later}
  \item Report bugs at \url{https://github.com/r-lib/later/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
  \item Winston Chang \email{winston@posit.co} (\href{https://orcid.org/0000-0002-1576-2126}{ORCID})
  \item Joe Cheng \email{joe@posit.co}
  \item Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR})
}

Other contributors:
\itemize{
  \item Marcus Geelnard (TinyCThread library, https://tinycthread.github.io/) [contributor, copyright holder]
  \item Evan Nemerson (TinyCThread library, https://tinycthread.github.io/) [contributor, copyright holder]
}

}
\keyword{internal}
