% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{later_dll_api_version}
\alias{later_dll_api_version}
\alias{later_h_api_version}
\alias{launchBgTask}
\alias{checkLaterOrdering}
\alias{cpp_error}
\alias{asyncFib}
\alias{testfd}
\title{Testing Function}
\usage{
later_dll_api_version()

later_h_api_version()

launchBgTask(secsToSleep)

checkLaterOrdering()

cpp_error(value)

asyncFib(resolve, reject, x)

testfd()
}
\arguments{
\item{secsToSleep}{seconds to sleep (integer)}

\item{value}{integer value}

\item{resolve}{object to resolve (sexp)}

\item{reject}{object to reject (sexp)}

\item{x}{value for Fibonnacci task (double)}
}
\description{
Testing Function

Testing Function

Testing Function

Testing Function

Testing Function

Testing Function

Testing Function
}
